move tests from datasets_test to crossval_test
